# docs/superpowers/plans/2026-06-17-api-spec.yaml
openapi: "3.0.3"
info:
  title: Intelligent Foreign Trade Financial API
  description: |
    API for the China foreign trade financial software (智能外贸财务系统).
    Features: AI-powered voucher generation, multi-ledger accounting, period closing,
    OEM commission calculation, print & export, audit trail.

    **Auth:** JWT Bearer token in `Authorization` header. CSRF via `X-CSRF-Token`.
    **Multi-ledger:** Pass `X-Ledger-Id` header to scope all operations to a ledger.
  version: "1.0.0"
  contact:
    name: Trad Account Team
servers:
  - url: http://localhost:8004
    description: Local development (8000-8003 blocked by zombie ports on Windows)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccountCreate:
      type: object
      required: [code, name, account_type, balance_direction]
      properties:
        code: { type: string, description: 科目编码, example: "1001" }
        name: { type: string, description: 科目名称, example: "库存现金" }
        parent_id: { type: integer, nullable: true }
        account_type: { type: string, enum: [ASSET, LIABILITY, EQUITY, COST, PROFIT_LOSS] }
        balance_direction: { type: string, enum: [DEBIT, CREDIT] }
        opening_balance: { type: number, default: 0.00 }
        is_active: { type: boolean, default: true }
    Account:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        code: { type: string }
        name: { type: string }
        parent_id: { type: integer, nullable: true }
        account_type: { type: string }
        balance_direction: { type: string }
        opening_balance: { type: number }
        is_active: { type: boolean }
        children: { type: array, items: { $ref: '#/components/schemas/Account' } }

    # --- Partner ---
    PartnerCreate:
      type: object
      required: [code, name, partner_type]
      properties:
        code: { type: string, description: 往来单位编码 }
        name: { type: string, description: 往来单位名称 }
        partner_type: { type: string, enum: [CUSTOMER, VENDOR, BOTH] }
        is_active: { type: boolean, default: true }
    Partner:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        code: { type: string }
        name: { type: string }
        partner_type: { type: string }
        is_active: { type: boolean }

    # --- Voucher ---
...
security:
  - BearerAuth: []
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List chart of accounts (hierarchical)
      parameters:
        - name: active_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Account' } } } } }
    post:
      tags: [Accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
  /accounts/{id}:
    put:
      tags: [Accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }

  # ========== PARTNERS ==========
  /partners:
    get:
      tags: [Partners]
      summary: List business partners
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [CUSTOMER, VENDOR, BOTH] }
        - name: active_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Partner' } } } } }
    post:
      tags: [Partners]
      summary: Create a business partner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Partner' } } } }

  # ========== VOUCHERS ==========
...